+++ src/types/debrief/index.d.ts
@@ -1,5 +1,5 @@
 import annotate, { annotateFields } from './annotate';
-import { isAnnotation, Annotation } from './ast';
+import { isAnnotation, Annotation } from './types';
 import serialize from './serialize';
 import summarize from './summarize';
 import { indent } from './utils';
